+++ Tests/ValidationSystem/Modules/Asleep/Asleep.js
@@ -1,0 +1,11 @@
+/**
+ * This is a comment to test that comments will not bereak
+ * class-only entities.
+ */
+class Asleep {
+
+	Ping(com, fun) {
+		fun(null, com);
+	}
+
+}