+++ src/Nexus.js
@@ -1,5 +1,9 @@
 global.pidInterchange = (pid) => { return { Value: pid, Format: 'is.xgraph.pid', toString: function () { return this.Value } } };
-module.exports = function xGraph() {
+
+const CacheInterface = require('./CacheInterface.js');
+
+module.exports = function xGraph(__options={}) {
+	this.__options=__options;
 	let eventListeners = {
 		exit: [],
 		setup: [],
@@ -17,34 +21,34 @@ module.exports = function xGraph() {
 		eventListeners[eventName].push(listener);
 	}
 
-	this.boot = function boot(optionsArray) {
-		(async function (__options) {
-			if (typeof state == "undefined") state = process.env.XGRAPH_ENV || "production";
-			if (__options.indexOf("--debug") > -1 || __options.indexOf("--development") > -1) {
-				state = 'development';
+	this.boot = function boot() {
+		return (async function (__options) {
+
+			function checkFlag(flag) {
+				// console.dir(__options);
+				return flag in __options && __options[flag];
 			}
+
 			// module.paths.push(process.cwd() + '/cache/node_modules');
 
 			console.log(`\nInitializing the Run Engine`);
 
 			const fs = require('fs');
 			const Path = require('path');
-			const jszip = require("jszip");
 			const endOfLine = require('os').EOL;
+
+			const jszip = require("jszip");
+			const Uuid = require('uuid/v4');
+			const stripComments = require('strip-comments');
+
 			var consoleNotification = false;
-			var Uuid;
-			var CacheDir;						// The location of where the Cache will be stored
+			let cacheInterface;
 			var Config = {};					// The read config.json
 			var ModCache = {};					// {<folder>: <module>}
-			var ApexIndex = {}; 				// {<Apex pid>:<folder>}
-			var Setup = {};						// {<Apex pid>:<function>}
-			var Start = {};						// {<Apex pid>:<function>}
 			var Stop = {};						// {<Apex pid>:<function>}
 			var EntCache = {};					// {<Entity pid>:<Entity>
 			var ImpCache = {};					// {<Implementation path>: <Implementation(e.g. disp)>}
 			var packagejson = {};				// The compiled package.json, built from Modules
-			var args = __options;			// The input __options --under consideration for deprication
-			var Params = {};					// The set of Macros for defining paths
 			var originalConsoleLog = console.log;
 			var Nxs = {
 				genPid,
@@ -97,6 +101,10 @@ module.exports = function xGraph() {
 				// The logging function for writing to xgraph.log to the current working directory
 				const xgraphlog = (...str) => {
 					xgraphlog.buffer.lock((val) => val + `${log.parse(str)}${endOfLine}`);
+					if(!xgraphlog.busy) {
+						xgraphlog.busy = true;
+						xgraphlog.updateInterval();
+					}
 				};
 				xgraphlog.buffer = new Volatile("");
 				xgraphlog.updateInterval = async () => {
@@ -106,10 +114,16 @@ module.exports = function xGraph() {
 						return "";
 					});
 					fs.appendFile(`${process.cwd()}/xgraph.log`, str, (err) => {
-						process.nextTick(xgraphlog.updateInterval);
+						xgraphlog.buffer.lock(val => {
+							if(val !== '') {
+								// we have more in out buffer, keep calling out to the thing
+								process.nextTick(xgraphlog.updateInterval);
+							}else {
+								xgraphlog.busy = false;
+							}
+						});
 					});
 				}
-				xgraphlog.updateInterval();
 
 
 				// The defined log levels for outputting to the std.out() (ex. log. v(), log. d() ...)
@@ -193,69 +207,58 @@ module.exports = function xGraph() {
 					console.timers[_] = undefined;
 					log.i(`${_}: ${elapsed}ms`);
 				}
-				process.on('unhandledRejection', event => {
-					log.e('------------------ [Stack] ------------------');
-					log.e(`line ${event.lineNumber}, ${event}`);
-					log.e(event.stack);
-					log.e('------------------ [/Stack] -----------------');
-					process.exit(1);
-				});
+				// process.on('unhandledRejection', event => {
+				// 	log.e('------------------ [Stack] ------------------');
+				// 	log.e(`line ${event.lineNumber}, ${event}`);
+				// 	log.e(event.stack);
+				// 	log.e('------------------ [/Stack] -----------------');
+				// 	process.exit(1);
+				// });
 
 			}
 
 			function indirectEvalImp(entString) {
+				//sanitize entString!
+				entString = stripComments(entString).trim();
+
 				let imp = (1, eval)(entString);
-				if(typeof imp != 'undefined') return imp;
-				else return { dispatch: ((1, eval)(`(function(){ return ${entString} })()`)).prototype };
+				if(typeof imp != 'undefined') {
+					if(!('dispatch' in imp)) {
+						log.e('Entity does not return a dispatch Table');
+						throw new Error('E_NO_DISPATCH_TABLE');
+					}
+					return imp;
+				}
+				else {
+					imp = { dispatch: ((1, eval)(`(function(){ return ${entString} })()`)).prototype };
+					if(typeof imp != 'undefined') {
+						if(!('dispatch' in imp)) {
+							log.e('Entity does not return a dispatch Table');
+							throw new Error('E_NO_DISPATCH_TABLE');
+						}
+						return imp;
+					}
+				}
 			}
 
 
 			log.i('=================================================');
 			log.i(`Nexus Warming Up:`);
 
-			defineMacros();
+			//set CWD
+			__options.cwd = __options.cwd ? Path.resolve(__options.cwd) : Path.resolve('.');
+			log.v(`CWD set to ${__options.cwd}`);
 
-			// if called from binary quit or if called from
-			// the command line and node build cache first
-			if (!fs.existsSync(CacheDir) || (state == "development")) {
-				// #ifndef BUILT
-				if (isBinary()) {
-					// #endif
-					log.e(`No cache exists at ${CacheDir}. Try xgraph run`);
-					process.exit(1);
-					return;
-					// #ifndef BUILT
-				}
-				log.i("Building the Cache");
-				log.i(__options[1]);
-				let genesisString = fs.readFileSync(`${__options[1].substr(0, __options[1].lastIndexOf(Path.sep))}/Genesis.js`).toString();
-				await eval(genesisString);
-				// #endif
+			//set Cache location
+			__options.cache = __options.cache || Path.join(__options.cwd, 'cache');
+
+			// if the cache doesnt exist, throw
+			if (!fs.existsSync(__options.cache)) {
+				log.e(`No cache exists at ${__options.cache}. Try xgraph run`);
+				throw new Error(`No cache exists at ${__options.cache}. Try xgraph run`);
 			}
 
 			initiate();
-
-			async function exit(code = 0) {
-				log.i(`--Nexus/Stop`);
-				//build the Stop promise array
-				let stopTasks = [];
-
-				for (let pid in Stop) {
-					stopTasks.push(new Promise((resolve, reject) => {
-						var com = {};
-						com.Cmd = Stop[pid];
-						com.Passport = {};
-						com.Passport.To = pid;
-						com.Passport.Pid = genPid();
-						sendMessage(com, resolve);
-					}));
-				}
-				console.log = originalConsoleLog;
-				await Promise.all(stopTasks);
-				log.v(`--Nexus: All Stops Complete`);
-
-				dispatchEvent('exit', { exitCode: code });
-			}
 
 			////////////////////////////////////////////////////////////////////////////////////////////////
 			//
@@ -263,53 +266,23 @@ module.exports = function xGraph() {
 			//
 			//
 
-			/**
-			 * Populates Params {OBJECT}
-			 * This is populated from both the __options array as well as those parsed in the
-			 * binary file if it was used.
-			 * Such asignments are of the form Config=... Cache=... or paths xGraph=....
-			 */
-			function defineMacros() {
-				// Process input __options and define macro parameters
-				// All macros are stored case insensitive in the Params object
-
-				let arg, parts;
-				for (var iarg = 0; iarg < args.length; iarg++) {
-					arg = args[iarg];
-					try {
-						let jarg = JSON.parse(arg);
-						for (let key in jarg) {
-							// log.v(`${key}=${jarg[key]}`);
-							Params[key] = jarg[key];
-						}
-					} catch (e) {
-						// log.v(arg);
-						parts = arg.split('=');
-						if (parts.length == 2) {
-							if (parts[1][0] != "/") parts[1] = Path.resolve(parts[1]);
-							Params[parts[0].toLowerCase()] = parts[1];
-						}
-					}
-				}
-
-				//set CWD
-				CWD = Params.cwd ? Path.resolve(Params.cwd) : Path.resolve('.');
-				log.v(`CWD set to ${CWD}`);
-
-				//set Cache location
-				CacheDir = Params.cache || Path.join(CWD, 'cache');
-			}
 
 			/**
 			 *  The main process of starting an xGraph System.
 			 */
 			async function initiate() {
 				log.i(`--Nexus/Initiate`);
-				ApexIndex = {};
-				Setup = {};
-				Start = {};
+				let Setup = {};
+				let Start = {};
+				cacheInterface = new CacheInterface({
+					path: __options.cache
+				});
 
-				loadCache();
+
+				let cache = await cacheInterface.loadCache();
+				Start = cache.start;
+				Setup = cache.setup;
+				Stop = Object.assign(Stop, cache.stop);
 
 				await setup();
 
@@ -323,52 +296,6 @@ module.exports = function xGraph() {
 				//
 				//
 
-				/**
-				 * Load in the cache and poulate setup Setup, Start, and ApexIndex {Objects}
-				 */
-				function loadCache() {
-					var folders = fs.readdirSync(CacheDir);
-
-					for (var ifold = 0; ifold < folders.length; ifold++) {
-						let folder = folders[ifold];
-						let path = `${CacheDir}/${folder}/Module.zip`;
-						if (!fs.existsSync(path))
-							continue;
-
-						parseMod(folder)
-
-						function parseMod(folder) {
-							let dir = CacheDir + '/' + folder;
-							var instancefiles = fs.readdirSync(dir);
-							for (var ifile = 0; ifile < instancefiles.length; ifile++) {
-								var file = instancefiles[ifile];
-
-								//check that it's an instance of the module
-								if (file.length !== 32)
-									continue;
-
-								var path = dir + '/' + file;
-								if (fs.lstatSync(path).isDirectory()) {
-									ApexIndex[file] = folder;
-									let instJson = JSON.parse(fs.readFileSync(`${path}/${file}.json`));
-
-									if ('$Setup' in instJson)
-										Setup[file] = instJson.$Setup;
-									if ('$Start' in instJson)
-										Start[file] = instJson.$Start;
-									if ('$Stop' in instJson)
-										Stop[file] = instJson.$Stop;
-
-									// log.d(file);
-								}
-							}
-						}
-					}
-
-					log.v('ApexIndex', JSON.stringify(ApexIndex, null, 2));
-					log.v('Setup', JSON.stringify(Setup, null, 2));
-					log.v('Start', JSON.stringify(Start, null, 2));
-				}
 
 				/**
 				 * Call setup on the required Module Apexes
@@ -434,16 +361,36 @@ module.exports = function xGraph() {
 			//
 			//
 
-
-			// #ifndef BUILT
 			/**
-			 * Check if the system is running from binary
+			 * @description stop the system by calling all stop methods and
+			 * releasing all entities from memory
+			 * exit code 0 will gracefully exit,
+			 * @param {number} [code=0] the code to exit with
 			 */
-			function isBinary() {
-				return (typeof tar !== 'undefined');
-			}
-			// #endif
+			async function exit(code = 0) {
+				log.i(`--Nexus/Stop`);
+				//build the Stop promise array
+				let stopTasks = [];
 
+				log.i('Nexus unloading node modules');
+				log.v(Object.keys(require.cache).join('\n'));
+
+				for (let pid in Stop) {
+					stopTasks.push(new Promise((resolve, reject) => {
+						var com = {};
+						com.Cmd = Stop[pid];
+						com.Passport = {};
+						com.Passport.To = pid;
+						com.Passport.Pid = genPid();
+						sendMessage(com, resolve);
+					}));
+				}
+				console.log = originalConsoleLog;
+				await Promise.all(stopTasks);
+				log.v(`--Nexus: All Stops Complete`);
+
+				dispatchEvent('exit', { exitCode: code });
+			}
 
 			/**
 			 * replace the macros for local path info
@@ -467,8 +414,8 @@ module.exports = function xGraph() {
 							break;
 						case 2:
 							if (chr == '}') {
-								if (param in Params)
-									s += Params[param];
+								if (param in __options)
+									s += __options[param];
 								else
 									throw 'Parameter <' + param + '> not defined';
 								state = 1;
@@ -488,10 +435,6 @@ module.exports = function xGraph() {
 			 * generate a 32 character hex pid
 			 */
 			function genPid() {
-				if (!Uuid) {
-					module.paths = [Path.join(Path.resolve(CacheDir), 'node_modules')];
-					Uuid = require('uuid/v4');
-				}
 				var str = Uuid();
 				var pid = str.replace(/-/g, '').toUpperCase();
 				return pid;
@@ -509,7 +452,6 @@ module.exports = function xGraph() {
 			 * @callback fun 				the callback function to return to when finished
 			 */
 			function sendMessage(com, fun = _ => _) {
-				// log.d('NEXUS MESSAGE:', com)
 				if (!('Passport' in com)) {
 					log.w('Message has no Passport, ignored');
 					log.w('    ' + JSON.stringify(com));
@@ -537,7 +479,8 @@ module.exports = function xGraph() {
 					done(null, EntCache[pid]);
 					return;
 				} else {
-					getEntityContext(apx, pid, done);
+					// if(pid in ApexIndex) apx = pid;
+					getEntityContext(pid, done);
 				}
 
 				function done(err, entContext) {
@@ -548,7 +491,7 @@ module.exports = function xGraph() {
 						return;
 					}
 
-					if ((pid in ApexIndex) || (entContext.Par.Apex == apx)) {
+					if ((pid in cacheInterface.ApexIndex) || (entContext.Par.Apex == apx)) {
 						entContext.dispatch(com, reply);
 					} else {
 						let err = 'Trying to send a message to a non-Apex'
@@ -595,7 +538,7 @@ module.exports = function xGraph() {
 				};
 
 				/**
-				 * load a dependency for a module
+				 * Given a module name, `require` loads the module, returning the module object.
 				 * @param {string} string 	the string of the module to require/load
 				 */
 				function require(string) {
@@ -603,7 +546,7 @@ module.exports = function xGraph() {
 				}
 
 				function exit(code) {
-					nxs.exit(code)
+					nxs.exit(code);
 				}
 
 				/**
@@ -644,15 +587,24 @@ module.exports = function xGraph() {
 				}
 
 				/**
-				 * entity access to the genModule command
-				 * @param {object} mod 	the description of the Module to generate
-				 * @param {string} mod.Module the module to generate
-				 * @param {object=} mod.Par 	the Par to merge with the modules Apex Par
+				 * Entity access to the genModule command.
+				 * genModule is the same as genModules.
+				 * genModule expects two parameters: moduleObject and fun.
+				 *
+				 * The moduleObject parameter is an object that contains data for each module that will be
+				 * generated. If only one module needs to be generated, then moduleObject can be a simple
+				 * module definition. If more then one module needs to be generated, moduleObject has a
+				 * key for each module definition, such as in a system structure object.
+				 *
+				 * When this.genModule is called from an entity, the moduleObject and fun parameters are passed
+				 * along to nxs.genModule, which starts the module and adds it to the system.
+				 * @param {object} moduleObject		Either a single module definition, or an object containing
+				 * 										multiple module definitions.
 				 * @callback fun
 				 */
-				function genModule(mod, fun) {
+				function genModule(moduleObject, fun) {
 					//	log.v('--Entity/genModule');
-					nxs.genModule(mod, fun);
+					nxs.genModule(moduleObject, fun);
 				}
 
 				/**
@@ -666,13 +618,23 @@ module.exports = function xGraph() {
 				}
 
 				/**
-				 * entity access to the genModule command
-				 * @param {object} modObj 	an object containing one or more module descriptions
-				 * @callback fun(err,pidofTop,objectOfAllModulesGenerated)
+				 * Entity access to the genModule command.
+				 * genModule expects two parameters: moduleObject and fun.
+				 *
+				 * The moduleObject parameter is an object that contains data for each module that will be
+				 * generated. If only one module needs to be generated, then moduleObject can be a simple
+				 * module definition. If more then one module needs to be generated, moduleObject has a
+				 * key for each module definition, such as in a system structure object.
+				 *
+				 * When this.genModule is called from an entity, the moduleObject and fun parameters are passed
+				 * along to nxs.genModule, which starts the module and adds it to the system.
+				 * @param {object} moduleObject		Either a single module definition, or an object containing
+				 * 										multiple module definitions.
+				 * @callback fun
 				 */
-				function genModules(modObj, fun) {
+				function genModules(moduleObject, fun) {
 					//	log.v('--Entity/genModule');
-					nxs.genModule(mod, fun);
+					nxs.genModule(moduleObject, fun);
 				}
 
 				/**
@@ -685,10 +647,10 @@ module.exports = function xGraph() {
 				}
 
 				/**
-				 * create an entity in the same module
-				 * @param {object} par the par of the entity to be generated
-				 * @param {string} par.Entity The entity type that will be generated
-				 * @param {string=} par.Pid	the pid to define as the pid of the entity
+				 * Create an entity in the same module. Entities can only communicate within a module.
+				 * @param {object} par 			The parameter object of the entity to be generated.
+				 * @param {string} par.Entity 	The entity type that will be generated.
+				 * @param {string=} par.Pid		The pid to set as the pid of the entity.
 				 * @callback fun
 				 */
 				function genEntity(par, fun) {
@@ -696,18 +658,18 @@ module.exports = function xGraph() {
 				}
 
 				/**
-				 * create a 32 character hexidecimal pid
+				 * Create and return a 32 character hexadecimal pid.
 				 */
 				function genPid() {
 					return nxs.genPid();
 				}
 
 				/**
-				 * Send a message to another entity, you can only send messages to Apexes of modules
-				 * unless both sender and recipient are in the same module
-				 * @param {object} com  		the message object to send
-				 * @param {string} com.Cmd		the function to send the message to in the destination entity
-				 * @param {string} pid 			the pid of the recipient (destination) entity
+				 * Sends the command object and the callback function to the xGraph part (entity or module, depending
+				 * on the fractal layer) specified in the Pid.
+				 * @param {object} com  	The message object to send.
+				 * @param {string} com.Cmd	The function to send the message to in the destination entity.
+				 * @param {string} pid 		The pid of the recipient (destination) entity.
 				 * @callback fun
 				 */
 				function send(com, pid, fun) {
@@ -725,8 +687,9 @@ module.exports = function xGraph() {
 				}
 
 				/**
-				 * save the current entity to cache if not an Apex send the save message to Apex
-				 * if it is an Apex we save it as well as all other relevant information
+				 * Save this entity, including it's current Par and Vlt, to the cache.
+				 * If this entity is not an Apex, send the save message to Apex of this entity's module.
+				 * If it is an Apex we save the entity's information, as well as all other relevant information
 				 * @callback fun
 				 */
 				function save(fun) {
@@ -735,13 +698,14 @@ module.exports = function xGraph() {
 			}
 
 			/**
-			 * Create an Entity from the given par in the module defined by apx
+			 * Creates an Entity in the module, which is defined by the apx, from the given entity definition
 			 * The entity is then stored in EntCache (the location of all "in Memory" entities)
-			 * @param {string} apx 		the Pid of the module Apex in which this entity will be generated
-			 * @param {object} par 		the Par of the entity that will be created
-			 * @param {string} par.Entity The entity type that will be generated
-			 * @param {string=} par.Pid	the pid to define as the pid of the entity
-			 * @callback fun 			the callback to return the pid of the generated entity to
+			 * @param {string} apx 			the Pid of the module Apex in which this entity will be generated
+			 * @param {object} par 			the Par of the entity that will be created
+			 * @param {string} par.Entity 	The entity type that will be generated
+			 * @param {string=} par.Pid		the pid to define as the pid of the entity
+			 * @return {pid} par.Pid		the pid of the generated entity
+			 * @callback fun
 			 */
 			async function genEntity(apx, par, fun = _ => log.e(_)) {
 				if (!("Entity" in par)) {
@@ -749,11 +713,11 @@ module.exports = function xGraph() {
 					return;
 				}
 
-				var impkey = ApexIndex[apx] + '/' + par.Entity;
-				var mod = ModCache[ApexIndex[apx]];
+				var impkey = Path.join(cacheInterface.ApexIndex[apx], par.Entity);
+				var mod = ModCache[cacheInterface.ApexIndex[apx]];
 
 				if (!(par.Entity in mod.files)) {
-					log.e('<' + par.Entity + '> not in module <' + ApexIndex[apx] + '>');
+					log.e('<' + par.Entity + '> not in module <' + cacheInterface.ApexIndex[apx] + '>');
 					fun('Null entity');
 					return;
 				}
@@ -766,7 +730,7 @@ module.exports = function xGraph() {
 				}
 
 				par.Pid = par.Pid || genPid();
-				par.Module = ApexIndex[apx];
+				par.Module = cacheInterface.ApexIndex[apx];
 				par.Apex = apx;
 
 				EntCache[par.Pid] = new Entity(Nxs, ImpCache[impkey], par);
@@ -774,14 +738,14 @@ module.exports = function xGraph() {
 			}
 
 			/**
-			 * Delete an entity file. If the entity is an Apex of a Module,
+			 * Delete an entity from the module's memory.  If the entity is an Apex of a Module,
 			 * then delete all the entities found in that module as well.
 			 * @param {string} apx 		the pid of the entities apex
 			 * @param {string} pid 		the pid of the entity
 			 * @callback fun  			the callback to return the pid of the generated entity to
 			 */
-			function deleteEntity(apx, pid, fun = _ => _) {
-				let apxpath = `${CacheDir}/${ApexIndex[apx]}/${apx}/`;
+			function deleteEntity(apx, pid, fun) {
+				let apxpath = `${__options.cache}/${cacheInterface.ApexIndex[apx]}/${apx}/`;
 
 				let rmList = [];
 				//we first check to see if it's an apex
@@ -805,7 +769,8 @@ module.exports = function xGraph() {
 						delete EntCache[subpid];
 					}
 				}
-				fun(null, pid);
+				if(fun)
+					fun(null, pid);
 			}
 
 			/**
@@ -816,7 +781,7 @@ module.exports = function xGraph() {
 			 * @callback fun  			the callback to return the pid of the generated entity to
 			 */
 			function saveEntity(apx, pid, fun = (err, pid) => { if (err) log.e(err) }) {
-				let modpath = `${CacheDir}/${ApexIndex[apx]}`;
+				let modpath = `${__options.cache}/${cacheInterface.ApexIndex[apx]}`;
 				let apxpath = `${modpath}/${apx}`;
 				let entpath = `${apxpath}/${pid}.json`;
 
@@ -830,7 +795,7 @@ module.exports = function xGraph() {
 							//the following code is deprecated since including deferred modules all module zip files
 							//must exist in the cache prior to starting the system
 
-							fun(`No Directory for the requested module: ${ApexIndex[apx]}`);
+							fun(`No Directory for the requested module: ${cacheInterface.ApexIndex[apx]}`);
 						}
 					})
 				});
@@ -848,12 +813,12 @@ module.exports = function xGraph() {
 								checkEntity();
 							} else {
 								var schema = await new Promise(async (res, rej) => {
-									if ('schema.json' in ModCache[ApexIndex[apx]]) {
-										ModCache[ApexIndex[apx]].file('schema.json').async('string').then(function (schemaString) {
+									if ('schema.json' in ModCache[cacheInterface.ApexIndex[apx]]) {
+										ModCache[cacheInterface.ApexIndex[apx]].file('schema.json').async('string').then(function (schemaString) {
 											res(JSON.parse(schemaString));
 										});
 									} else {
-										log.e('Module <' + ApexIndex[apx] + '> schema not in ModCache');
+										log.e('Module <' + cacheInterface.ApexIndex[apx] + '> schema not in ModCache');
 										res();
 										return;
 									}
@@ -934,19 +899,9 @@ module.exports = function xGraph() {
 			 * @param {string} str 			the string of the module to require
 			 */
 			function loadDependency(apx, pid, str) {
-				// log.d(`Nexus::loadDependency (${apx}, ${pid}, ${str})`);
 
-				let folder = ApexIndex[apx];
-				try {
-					return require(str);
-				} catch (e) {
-					try {
-						return require(Path.join(CacheDir, folder, 'node_modules', str));
-					} catch (e) {
-						log.e(e);
-						process.exit(1);
-					}
-				}
+				let moduleType = cacheInterface.ApexIndex[apx];
+				return cacheInterface.loadDependency(moduleType, str)
 			}
 
 			/**
@@ -956,42 +911,35 @@ module.exports = function xGraph() {
 			 * @param {string} pid 		the pid of the entity
 			 * @callback fun  			the callback to return the pid of the generated entity to
 			 */
-			function getEntityContext(apx, pid, fun = _ => _) {
+			function getEntityContext(pid, fun = _ => _) {
 				let imp;
 				let par;
 				let ent;
 
-				// Check to see if pid is also an apex entity in this system
-				// if not then we assume that the pid is an entity inside of the sending Module
-				if (pid in ApexIndex) {
-					apx = pid;
-				}
-
-				let folder = ApexIndex[apx];
-				let path = CacheDir + '/' + folder + '/' + apx + '/' + pid + '.json';
-				fs.readFile(path, function (err, data) {
+				cacheInterface.getEntityPar(pid, (err, data) => {
 					if (err) {
-						log.e('<' + path + '> unavailable');
+						log.e(`Error retrieving a ${moduleType} from cache. Pid: ${pid}`);
+						log.e(err);
 						fun('Unavailable');
 						return;
 					}
 					let par = JSON.parse(data.toString());
-					let impkey = folder + '/' + par.Entity;
+					let impkey = par.Module + '/' + par.Entity;
 					let imp;
 					if (impkey in ImpCache) {
 						BuildEnt();
 						return;
 					}
 
-					GetModule(folder, async function (err, mod) {
+					GetModule(par.Module, async function (err, mod) {
 						if (err) {
-							log.e('Module <' + folder + '> not available');
+							log.e('Module <' + par.Module + '> not available');
 							fun('Module not available');
 							return;
 						}
 
 						if (!(par.Entity in mod.files)) {
-							log.e('<' + par.Entity + '> not in module <' + folder + '>');
+							log.e('<' + par.Entity + '> not in module <' + par.Module + '>');
 							fun('Null entity');
 							return;
 						}
@@ -1000,12 +948,13 @@ module.exports = function xGraph() {
 							mod.file(par.Entity).async("string").then((string) => res(string))
 						});
 
-						log.v(`Spinning up entity ${folder}-${par.Entity.split('.')[0]}`);
+						log.v(`Spinning up entity ${par.Module}-${par.Entity.split('.')[0]}`);
 						ImpCache[impkey] = indirectEvalImp(entString);
 						BuildEnt();
 					});
 
 					function BuildEnt() {
+						// TODO: rethink the whole process of having to call out a setup and start
 						EntCache[pid] = new Entity(Nxs, ImpCache[impkey], par);
 						fun(null, EntCache[pid]);
 					}
@@ -1050,7 +999,7 @@ module.exports = function xGraph() {
 								return;
 							}
 							let pidapx = symbols[moduleKey];
-							ApexIndex[pidapx] = inst.Module;
+							cacheInterface.ApexIndex[pidapx] = inst.Module;
 
 
 							for (let key in inst.Par) {
@@ -1063,8 +1012,6 @@ module.exports = function xGraph() {
 											inst.Par[key] = symbols[symbol];
 										} else {
 											log.w(`${symbol} not in Module key list`);
-											log.d(symbol);
-											log.d(symbols);
 											log.v(`${Object.keys(symbols)}`)
 										}
 									}
@@ -1298,41 +1245,14 @@ module.exports = function xGraph() {
 			 */
 			function GetModule(ModName, fun = _ => _) {
 				ModName = ModName.replace(/\:\//g, '.');
-
-				//get the module from memory (ModCache) if it has already been retrieved
 				if (ModName in ModCache) return fun(null, ModCache[ModName]);
-
-				//get the module from cache
-				var cachedMod = `${CacheDir}/${ModName}/Module.zip`;
-				fs.lstat(cachedMod, function (err, stat) {
-					if (err) {
-						log.e(`Error retreiving ${cachedMod} from cache`);
-						fun(err);
-					}
-					if (stat) {
-						if (!stat.isDirectory()) {
-							fs.readFile(cachedMod, async function (err, data) {
-								if (err) {
-									fun(err);
-									return;
-								}
-								ModCache[ModName] = await new Promise(async (res, rej) => {
-									let zip = new jszip();
-									zip.loadAsync(data).then((mod) => res(mod));
-								});
-								fun(null, ModCache[ModName]);
-								return;
-							});
-						}
-					} else {
-						err = `Module ${cachedMod} does not exist in the cache`
-						log.e(err);
-						fun(err);
-						return;
-					}
+				else cacheInterface.getModule(ModName, (err, moduleZip) => {
+					if(err) return fun(err);
+					ModCache[ModName] = moduleZip;
+					return fun(null, ModCache[ModName]);
 				});
 			}
-		})(optionsArray);
+		})(this.__options);
 	}
 }